refactor: some code refactoring, updated FormBuilder to reuse code from public header for computation of design constants
