+++ frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignDrawer/EditFieldDrawer/EditFieldDrawer.tsx
@@ -5,7 +5,7 @@ import { BasicField, FieldCreateDto } from '~shared/types/field'
 import { BASICFIELD_TO_DRAWER_META } from '~features/admin-form/create/constants'
 import { isMyInfo } from '~features/myinfo/utils'
 
-import { useBuilderFields } from '../../BuilderAndDesignContent/BuilderFields'
+import { useBuilderFields } from '../../BuilderAndDesignContent/useBuilderFields'
 import { MYINFO_FIELD_CONSTANTS } from '../../constants'
 import {
   BuildFieldState,
