+++ frontend/src/features/admin-form/create/builder-and-design/useDesignStore.tsx
@@ -2,25 +2,26 @@ import { isEqual } from 'lodash'
 import create from 'zustand'
 import { devtools } from 'zustand/middleware'
 
-import { FormColorTheme, FormLogoState, FormStartPage } from '~shared/types'
+import { FormStartPage } from '~shared/types'
 
 export type DesignStore = {
-  startPageData: FormStartPage
-  setState: (startPage: FormStartPage) => void
+  startPageData: FormStartPage | null
+  setStartPageData: (startPage: FormStartPage) => void
+  resetStartPageData: () => void
 }
 
 export const useDesignStore = create<DesignStore>(
   devtools((set, get) => ({
-    startPageData: {
-      logo: { state: FormLogoState.Default },
-      colorTheme: FormColorTheme.Blue,
-    },
-    setState: (startPage) => {
+    startPageData: null,
+    setStartPageData: (startPage) => {
       const current = get()
       if (isEqual(current.startPageData, startPage)) return
       set({
         startPageData: startPage,
       })
+    },
+    resetStartPageData: () => {
+      set({ startPageData: null })
     },
   })),
 )
@@ -29,5 +30,10 @@ export const startPageDataSelector = (
   state: DesignStore,
 ): DesignStore['startPageData'] => state.startPageData
 
-export const setStateSelector = (state: DesignStore): DesignStore['setState'] =>
-  state.setState
+export const setStartPageDataSelector = (
+  state: DesignStore,
+): DesignStore['setStartPageData'] => state.setStartPageData
+
+export const resetStartPageDataSelector = (
+  state: DesignStore,
+): DesignStore['resetStartPageData'] => state.resetStartPageData
